Add a comment for later consideration.
When clipping the Voronoi diagram, the buffered convex hull of every
centroids is used. Although it seems fine with tightly positionned
vertices, immense Voronoi cells are inevitably going go appear again
with some sparser configurations (but in the middle of the diagram
this time). To avoid this, we need to use a concave hull and consider
hollow part of the city.