Add the possibility to dynamically insert new lots
To avoid recomputing the complete Voronoi diagram, a sub-diagram is
generated based on the new lot, the "old" lot (which is going to be
divided) and the surrounding lots. The individual cells of this new
sub-diagram are then clipped with their previous form to avoid
overlays when merging with the global diagram.

TODO:
- update neighborhood relationships
- update the road network