Fix a special case when inserting a lot
After having inserted the new lot and computed the new polygons of the
surrounding lots, the intersection operation of the clipping process
often returned a GeometryCollection (containing several Geometry)
instead of a single Geometry. The origin of this phenomenom is unknown
but I suspect a floating precision bug.

To fix this, we check for this special case and only keep the Polygon
from the GeometryCollection, ignoring every other type of Geometry.