Fix wrong cliques being considered crossroads
In some case, a triangular Voronoi cell is surrounded by three other
cells (typically one by side). They are all neighbors so they form a
clique but they don't share a common vertex! For the moment this kind
of configuration is ignored to avoid bug but I fear the Bron-Kerbosh
algorithm is ill-adapted.