Finally, roads from shapefiles are loaded!
After a variety of different approaches, I finally got this
working. It seems that JTS does not scale very well to complex
multipolygon, especially when dealing with intersection tests.

The solution was to break the ginormous road network polygon into
small sections (in the form of LineStrings) and them perform the
intersection tests separately.